Made sure the sidebar always shows something
Now that CodeFund is gone, posts with the Other tag had an empty sidebar. This update just chooses one of the language tags and uses that to display language articles, otherwise it defaults to HTML. More thought should be put into what shows in the sidebar for articles like these in future.