FINALLY FIXED SIDEBAR FLASHING!!!
In the end, the issue was that the page was already being rendered before the sidebar HTML had been received, but after the article HTML had started being received. This meant that the CSS rendering engine was unaware of the sidebar's existence while it was trying to initially render the post content. I managed to figure this out by adding 100,000 "hello"s to the top of the article content, then throttling the network on the maximum setting. You could see the hellos getting loaded, then as soon as it was done loading all the article content, it would load the sidebar and the layout would shift into the correct positions. I fixed it by including the sidebar HTML (which is pretty small) _before_ the article HTML, then using flex-direction: row-reverse to make up for it.